backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI Backend for Todo App"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "asyncpg",
    "pydantic-settings",
    "pyjwt",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
